feat(inheritance-js-implementation): organising classes into files
