+++ src/classes/vehicle.js
@@ -1,0 +1,9 @@
+export class Vehicle{
+    start(){
+        console.log('starting Vehicle');
+    }
+
+    static getCompanyName(){
+        console.log('My Company')
+    }
+}
